requirements.txt
python-dotenv
google-generativeai
streamlit
sentence-transformers
numpy
pandas